+++ test/libsolidity/syntaxTests/immutable/base_ctor_return_no_immutables.sol
@@ -1,0 +1,10 @@
+contract Parent {
+    constructor() {
+        return;
+    }
+}
+
+contract Child is Parent {
+    uint public immutable baked = 123;
+}
+