+++ Changelog.md
@@ -11,6 +11,7 @@ Compiler Features:
 
 Bugfixes:
  * Control Flow Graph: Perform proper virtual lookup for modifiers for uninitialized variable and unreachable code analysis.
+ * Immutables: Fix wrong error when the constructor of a base contract uses ``return`` and the parent contract contains immutable variables.
 
 
 Solc-Js: